added option for logarithmic axes
Log axes can now be plotted with the logaxis option, as in:
plot(hexbin(y~x), logaxis=T)
The code for logarithmic axis in grid graphics comes straight from the
hydrosanity package, once on CRAN but now removed from it, and only available
on github. Hydrosanity is licensed as GPL>=2, which is compatible with
hexbin which uses GPL==2.